Enforce semver resolutions due to breaking changes
- There is a semver CVE that the ckeditor dev utils team has not addressed until later versions.
However, upgrading dev utils leads to breaking changes,
meaning we need to enforce semver resolutions in package.json in order
to address the CVE for ckeditor utils.

This package only applies to dev dependencies, and the change is strictly due to being unable to update ckeditor dev utils

This can be removed when we update ckeditor, but that is a large piece of work and currently not in scope to address this out of SLA vulnerability.