Trigger a new release to use updated license keys
With the migration to the latest version of CKEditor, we
need to update our license key to use one that is compatible
with the new version.

I updated the license key in the GitHub repository's secrets,
and we need to trigger a new release to ensure that the package
is built with the correct key, so this change just adds a
new changeset to create a release.